Migrate Embeddings Generation Service To TS
