+++ src/services/emb_management_service.js
@@ -3,7 +3,7 @@
 const {EmbeddingsStatus} = require("../model/embeddings_status");
 const {pruneEmbeddings, getCountDocumentsWithoutEmbedding} = require("./emb_collections_service");
 const {updateEmbeddingsStatusDict} = require("./emb_status_service");
-const {generateBatchesForModel} = require("../services/emb_generation_service");
+const {generateBatchesForModel} = require("./emb_generation_service");
 const {embeddingsTargetsAreValid} = require("../utils/embeddings_target");
 
 function pruneEmbeddingsIfNeeded(embeddingsStatusDict, overwriteExisting) {
